Expanded regex to reject similar but invalid file extensions, moved validation tests to test_serializers
