+++ tdrs-backend/tdpservice/reports/test/test_serializers.py
@@ -1,8 +1,10 @@
 """Test report serializers."""
-
+from django.core.exceptions import ValidationError
 import pytest
-from ..serializers import ReportFileSerializer
-from ..errors import ImmutabilityError
+
+from tdpservice.reports.errors import ImmutabilityError
+from tdpservice.reports.serializers import ReportFileSerializer
+from tdpservice.reports.validators import validate_file_extension
 
 
 @pytest.mark.django_db
@@ -30,19 +32,16 @@ def test_serializer_increment_create(report_data, other_report_data):
 @pytest.mark.django_db
 def test_immutability_of_report(report):
     """Test that report can only be created."""
-    try:
+    with pytest.raises(ImmutabilityError):
         serializer = ReportFileSerializer(
             report, data={
                 "original_filename": "BadGuy.js"
             },
-            partial=True)
+            partial=True
+        )
 
         serializer.is_valid()
         serializer.save()
-        raise Exception("Should not be able to update reports.")
-    except ImmutabilityError as err:
-        expected = "Cannot update, reports are immutable. Create a new one instead."
-        assert str(err) == expected
 
 
 @pytest.mark.django_db
@@ -53,3 +52,38 @@ def test_created_at(report_data):
     report = create_serializer.save()
 
     assert report.created_at
+
+
+@pytest.mark.parametrize("file_name", [
+    'sample.txt',
+    'Sample',
+    'ADS.E2J.FTP4.TS06.txt',
+    'ADS.E2J.FTP4.TS06',
+    'ADS.E2J.FTP2.MS18',
+    'ADS.E2J.FTP1.TS278'
+])
+def test_accepts_valid_file_extensions(file_name):
+    """Test valid file names are accepted by serializer validation."""
+    try:
+        validate_file_extension(file_name)
+    except ValidationError as err:
+        pytest.fail(f'Received unexpected error: {err}')
+
+
+@pytest.mark.parametrize("file_name", [
+    'java.jar',
+    'mysql.bin',
+    'malicious.exe',
+    'exec.py',
+    'hax.pdf',
+    'types.ts',
+    'notepad.txt.exe',
+    'ADS.E2J.FTP1.TS273894',
+    'ADS.E2J.FTP1.TS38WRONG',
+    'ADS.E2J.FTP1.MS38483',
+    'ADS.E2J.FTP1.MS38WRONG'
+])
+def test_rejects_invalid_file_extensions(file_name):
+    """Test invalid file names are rejected by serializer validation."""
+    with pytest.raises(ValidationError):
+        validate_file_extension(file_name)
